pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "visual-layer-mcp-server"
version = "1.0.2"
description = "MCP Server for Visual Layer SDK"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Visual Layer", email = "support@visual-layer.com"}
]
maintainers = [
    {name = "Visual Layer", email = "support@visual-layer.com"}
]
keywords = ["mcp", "visual-layer", "dataset", "ai", "machine-learning"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "pathlib>=1.0.1",
    "python-dotenv>=1.0.0",
    "PyJWT>=2.8.0",
    "pandas>=2.2.0",
    "mcp>=1.12.4",
    "fastmcp>=2.11.3",
    "visual-layer-sdk",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "typeguard",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
visual-layer-mcp = "mcp_server.server:main"

[project.urls]
Homepage = "https://github.com/visual-layer/vl-sdk"
Documentation = "https://docs.visual-layer.com"
Repository = "https://github.com/visual-layer/vl-sdk"
Issues = "https://github.com/visual-layer/vl-sdk/issues"

[tool.hatch.build.targets.wheel]
packages = ["mcp_server"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["mcp_server"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.uv]
dev-dependencies = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "typeguard",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
extra-index-url = ["https://test.pypi.org/simple"]

 